pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "meme-indexer"
version = "0.1.0"
description = "Persistent filesystem meme indexer with OCR-backed SQLite search."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=10.4.0",
  "pytesseract>=0.3.13",
  "tqdm>=4.66.0",
]

[tool.setuptools]
packages = ["meme_indexer"]
